requirements.txt
Babel==2.12.1
click==8.1.3
Flask==2.2.3
flask-babel==3.1.0
Flask-Table==0.5.0
gunicorn==20.1.0
importlib-metadata==6.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.5.0
pytz==2023.3
typing-extensions==4.5.0
Werkzeug==2.2.3
zipp==3.15.0
